Update 01.Responsive Navbar with Login on right align.html
As we are making the responsive navbar, the nav-items will align side by side. To align the nav-items like Register and login we use the mr-auto class in first <ul> tag and ml-auto in second <ul> tag. Such that the nav-items present in the second <ul> tag move to the right corner.